# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Avalanche P-Chain and EVM Indexer API\n\n## Overview\nThis API provides access to indexed Avalanche blockchain data including:\n- **Data API** (`/api/v1/data/*`): Blocks, transactions, subnets, validators, P-chain data\n- **Metrics API** (`/api/v1/metrics/*`): Fee statistics, chain metrics, time series data\n\n## Rate Limiting\nAll endpoints are rate limited to 60 requests/minute per IP with a burst of 10.",
        "title": "Icicle API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/v1/data/evm/{chainId}/address/{address}/txs": {
            "get": {
                "description": "Get transactions sent from or to a specific address",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Data - EVM"
                ],
                "summary": "Get transactions by address",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Chain ID",
                        "name": "chainId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Wallet address (0x-prefixed)",
                        "name": "address",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Number of results (max 100)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Pagination offset",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/api.Transaction"
                                            }
                                        },
                                        "meta": {
                                            "$ref": "#/definitions/api.Meta"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/v1/data/evm/{chainId}/txs": {
            "get": {
                "description": "Get a paginated list of transactions for a specific chain",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Data - EVM"
                ],
                "summary": "List transactions",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Chain ID",
                        "name": "chainId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Number of results (max 100)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Pagination offset",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/api.Transaction"
                                            }
                                        },
                                        "meta": {
                                            "$ref": "#/definitions/api.Meta"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/data/evm/{chainId}/txs/{hash}": {
            "get": {
                "description": "Get details for a specific transaction",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Data - EVM"
                ],
                "summary": "Get transaction by hash",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Chain ID",
                        "name": "chainId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Transaction hash (0x-prefixed)",
                        "name": "hash",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/api.Transaction"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "api.APIError": {
            "type": "object",
            "properties": {
                "code": {
                    "$ref": "#/definitions/api.ErrorCode"
                },
                "details": {
                    "type": "string"
                },
                "fields": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.FieldError"
                    }
                },
                "message": {
                    "type": "string"
                },
                "retry_after": {
                    "type": "integer"
                }
            }
        },
...
        "api.ErrorCode": {
            "type": "string",
            "enum": [
                "INVALID_PARAMETER",
                "NOT_FOUND",
                "INTERNAL_ERROR",
                "RATE_LIMITED",
                "VALIDATION_FAILED",
                "DATABASE_ERROR"
            ],
            "x-enum-varnames": [
                "ErrInvalidParameter",
                "ErrNotFound",
                "ErrInternalError",
                "ErrRateLimited",
                "ErrValidationFailed",
                "ErrDatabaseError"
            ]
        },
        "api.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/api.APIError"
                }
            }
        },
...
        "api.FieldError": {
            "type": "object",
            "properties": {
                "field": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "api.Meta": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "offset": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "api.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "meta": {
                    "$ref": "#/definitions/api.Meta"
                }
            }
        },
...
        "api.Transaction": {
            "type": "object",
            "properties": {
                "block_number": {
                    "type": "integer",
                    "example": 12345678
                },
                "block_time": {
                    "type": "string"
                },
                "chain_id": {
                    "type": "integer",
                    "example": 43114
                },
                "from": {
                    "type": "string",
                    "example": "0x742d35Cc6634C0532925a3b844Bc9e7595f..."
                },
                "gas_limit": {
                    "type": "integer",
                    "example": 21000
                },
                "gas_price": {
                    "type": "integer",
                    "example": 25000000000
                },
                "gas_used": {
                    "type": "integer",
                    "example": 21000
                },
                "hash": {
                    "type": "string",
                    "example": "0x1234..."
                },
                "success": {
                    "type": "boolean",
                    "example": true
                },
                "to": {
                    "description": "null for contract creation",
                    "type": "string",
                    "example": "0x123..."
                },
                "transaction_index": {
                    "type": "integer",
                    "example": 0
                },
                "type": {
                    "type": "integer",
                    "example": 2
                },
                "value": {
                    "type": "string",
                    "example": "1000000000000000000"
                }
            }
        },
...
    },
...
}